refactor: Move initialize, make_factories, use_factories, from_csv to StatTracker and factory attributes to StatTracker from StatTrackerCalculator
